requirements.txt
Flask==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
google-generativeai==0.3.2
APScheduler==3.10.4
python-dotenv==1.0.0
lxml==4.9.3
pytz==2023.3
gunicorn==21.2.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
cachetools==5.3.2